* Converted commodities back to int.

+++ src/pilot.h
@@ -138,7 +138,7 @@ typedef struct PilotOutfitSlot_ {
  */
 typedef struct PilotCommodity_ {
    Commodity* commodity; /**< Assosciated commodity. */
-   double quantity; /**< Amount player has. */
+   int quantity; /**< Amount player has. */
    unsigned int id; /**< Special mission id for cargo, 0 means none. */
 } PilotCommodity;
 
@@ -245,7 +245,7 @@ typedef struct Pilot_ {
    unsigned int credits; /**< monies the pilot has */
    PilotCommodity* commodities; /**< commodity and quantity */
    int ncommodities; /**< number of commodities. */
-   double cargo_free; /**< Free commodity space. */
+   int cargo_free; /**< Free commodity space. */
 
    /* Weapon properties */
    double weap_range; /**< Average range of primary weapons */
@@ -325,13 +325,13 @@ void pilot_switchSecondary( Pilot* p, PilotOutfitSlot *w );
  * Cargo.
  */
 /* Normal. */
-double pilot_cargoUsed( Pilot* pilot ); /* gets how much cargo it has onboard */
-double pilot_cargoFree( Pilot* p ); /* cargo space */
-double pilot_addCargo( Pilot* pilot, Commodity* cargo, double quantity );
-double pilot_rmCargo( Pilot* pilot, Commodity* cargo, double quantity );
+int pilot_cargoUsed( Pilot* pilot ); /* gets how much cargo it has onboard */
+int pilot_cargoFree( Pilot* p ); /* cargo space */
+int pilot_addCargo( Pilot* pilot, Commodity* cargo, int quantity );
+int pilot_rmCargo( Pilot* pilot, Commodity* cargo, int quantity );
 int pilot_moveCargo( Pilot* dest, Pilot* src );
 /* mission cargo - not to be confused with normal cargo */
-unsigned int pilot_addMissionCargo( Pilot* pilot, Commodity* cargo, double quantity );
+unsigned int pilot_addMissionCargo( Pilot* pilot, Commodity* cargo, int quantity );
 int pilot_rmMissionCargo( Pilot* pilot, unsigned int cargo_id, int jettison );
